+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- traintuple can take aggregatetuple as in_tuple
 - Pretty print for Model and Spec
 - Display orchestrator and chaincode versions in node info command
 - Accept full datamanager, metrics and parent_tasks objects in get_*_tuple responses